+++ include/mvh/interrupt.h
@@ -4,5 +4,6 @@
 void interrupt_init(void);
 void interrupt_enable(void);
 void interrupt_disable(void);
+void exception_dispatch(void *frame) __attribute__((noreturn));
 
 #endif