+++ include/mvh/panic.h
@@ -1,0 +1,32 @@
+#ifndef MVH_PANIC_H
+#define MVH_PANIC_H
+
+#include <stdint.h>
+
+typedef struct {
+    uint64_t r15;
+    uint64_t r14;
+    uint64_t r13;
+    uint64_t r12;
+    uint64_t r11;
+    uint64_t r10;
+    uint64_t r9;
+    uint64_t r8;
+    uint64_t rbp;
+    uint64_t rdi;
+    uint64_t rsi;
+    uint64_t rdx;
+    uint64_t rcx;
+    uint64_t rbx;
+    uint64_t rax;
+    uint64_t vector;
+    uint64_t error_code;
+    uint64_t rip;
+    uint64_t cs;
+    uint64_t rflags;
+} exception_frame_t;
+
+void kernel_panic(const char *message) __attribute__((noreturn));
+void kernel_panic_exception(const exception_frame_t *frame) __attribute__((noreturn));
+
+#endif